+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to the "commit-message-editor" extension will be documented in this file.
 
+## [Unreleased]
+
+### Fixed
+
+- Snycronize repository selector instances.
+
 ## [0.24.2] - 2023-05-08
 
 ### Fixed